pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "langfuse>=2.60.4",
    "langgraph>=0.4.1",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.5",
    "ollama>=0.4.8",
    "pandas>=2.2.3",
    "polars>=1.28.1",
    "pydantic-settings>=2.9.1",
    "rich>=13.9.4",
    "pygraphviz>=1.14",
    "openai==1.76.0",
    "instructor[groq]>=1.8.1",
    "langgraph-prebuilt>=0.1.8",
    "langgraph-sdk>=0.1.66",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "langsmith>=0.3.42",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langgraph-cli[inmem]>=0.2.10",
    "spacy>=3.8.5",
    "pip>=25.1.1",
    "jinja2>=3.1.6",
    "tavily-python>=0.7.2",
    "langchain-mistralai>=0.2.10",
    "langchain-tavily>=0.1.6",
    "torch==2.2.0",
    "outlines>=0.2.3",
    "langchain-litellm>=0.2.1",
    "tenacity>=9.1.2",
    "transformers>=4.52.3",
    "sqlalchemy>=2.0.40",
    "wikipedia>=1.4.0",
    "langchain-google-genai>=2.1.4",
    "trustcall>=0.0.39",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nb-black-formatter>=1.0.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.6",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
